requirements.txt
argcomplete==1.12.3
attrs==21.2.0
boto3==1.19.12
botocore==1.22.12
c7n==0.9.14
docutils==0.17.1
importlib-metadata==4.8.1
jmespath==0.10.0
jsonschema==3.2.0
pyrsistent==0.18.0
python-dateutil==2.8.2
PyYAML==5.4.1
s3transfer==0.5.0
six==1.16.0
tabulate==0.8.9
typing-extensions==3.10.0.2
urllib3==1.26.7
zipp==3.6.0
